# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ISetDeviceInfoResponse": {
				"type": "boolean"
			},
			"Error": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISetDeviceInfoBody": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"version": {
						"type": "string",
						"nullable": true
					},
					"baseOs": {
						"type": "string",
						"enum": [
							"ios",
							"android",
							"windows",
							"web"
						]
					}
				},
				"required": [
					"userId",
					"version",
					"baseOs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateStatBody": {
				"properties": {
					"statId": {
						"type": "number",
						"format": "double"
					},
					"newValue": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"statId",
					"newValue"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateNewStatBody": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"type",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDeleteStatBody": {
				"properties": {
					"statId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"statId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "typescript-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/v1/stats/deviceinfo": {
			"post": {
				"operationId": "SetDeviceInfo",
				"responses": {
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISetDeviceInfoResponse"
										},
										{
											"$ref": "#/components/schemas/Error"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Statistics"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISetDeviceInfoBody"
							}
						}
					}
				}
			}
		},
		"/v1/stats": {
			"get": {
				"operationId": "GetAllStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Statistics"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateStat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Statistics"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateStatBody"
							}
						}
					}
				}
			},
			"post": {
				"operationId": "CreateNewStat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Statistics"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateNewStatBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Statistics"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDeleteStatBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}